requirements.txt
# =============================================================================
# Sistema de Gestión de Crisis IPR — GORE Ñuble
# requirements.txt
# =============================================================================

# Framework
Flask==3.0.0
Werkzeug==3.0.1

# ORM
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Extensiones Flask
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1

# Servidor WSGI
gunicorn==21.2.0

# Utilidades
python-dotenv==1.0.0
email-validator==2.1.0
